requirements.txt
torch
numpy
opencv-python
Pillow
scipy
tqdm
gradio==3.17.0
transformers
